Fix the armor not being stored.
Do note that after looking over this code, I have become concerned as
when the player is jailed without a chest in the cell their inventory is
only stored temporary in the server's memory and if the server was to
crash or shutdown their inventory would be lost. That is why I want to
focus on the complete rewriting of the plugin, as the previous
maintainer removed all of this functionality for the temp storage.