Alright, may have found the cause of all the errors.
What was happening is the PlayerTeleport event was causing a player move
event and on the player move event we was then teleporting elsewhere
thus causing another PlayerTeleportEvent and it was an infinite loop.
Moved over the hand cuff listeners to their own file, see if this fixes
the problem now.